LayoutTests/editing/execCommand/indent-nested-lists-2-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of OL > DIV > BODY > HTML > #document to 9 of OL > DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:range from 4 of #text > LI > OL > DIV > BODY > HTML > #document to 4 of #text > LI > OL > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This tests indenting "three". You should see numbers 1 and 2 before "two" and "three" respectively.
one
two
three
four
Before: <li>one</li> <ol><li>two</li></ol> <li id="test">three</li> <li>four</li>
After: <li>one</li> <ol><li>two</li><li id="test">three</li></ol> <li>four</li>
